fix(upgrade): enable for all platforms
resolves #5184
