+++ src/upgrade/notice.go
@@ -32,6 +32,13 @@ type release struct {
 const (
 	RELEASEURL = "https://api.github.com/repos/jandedobbeleer/oh-my-posh/releases/latest"
 	CACHEKEY   = "upgrade_check"
+
+	upgradeNotice = `
+A new release of Oh My Posh is available: %s → %s
+To upgrade, run: 'oh-my-posh upgrade'
+
+To enable automated upgrades, set 'auto_upgrade' to 'true' in your configuration.
+`
 )
 
 func Latest(env platform.Environment) (string, error) {
