+++ website/docs/configuration/general.mdx
@@ -137,7 +137,7 @@ For example, the following is a valid `--config` flag:
 | `disable_cursor_positioning` | `boolean`        | disable fetching the cursor position in bash and zsh in case of unwanted side-effects                                                                                                                                                                                        |
 | `patch_pwsh_bleed`           | `boolean`        | patch a PowerShell bug where the background colors bleed into the next line at the end of the buffer (can be removed when [this][pwsh-bleed] is merged)                                                                                                                      |
 | `disable_notice`             | `boolean`        | disables the notice that a new upgrade is available when `auto_upgrade` is disabled, and the notice that no new upgrade is available when `auto_upgrade` is enabled                                                                                                          |
-| `auto_upgrade`               | `boolean`        | enable automatic upgrades for Oh My Posh (supports Windows/macOS only)                                                                                                                                                                                                       |
+| `auto_upgrade`               | `boolean`        | enable automatic upgrades for Oh My Posh (supports Windows, macOS and Linux)                                                                                                                                                                                                 |
 | `iterm_features`             | `[]string`       | enable iTerm2 specific features:<ul><li>`prompt_mark`: add the `iterm2_prompt_mark` [function][iterm2-si] for supported shells</li><li>`current_dir`: expose the current directory for iTerm2</li><li>`remote_host`: expose the current remote and user for iTerm2</li></ul> |
 
 ### JSON Schema Validation
